# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Thanarah Presentation Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResult"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user directly
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/disable:
    post:
      operationId: disableUser
      tags: [users]
      summary: Disable a user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{id}/force-logout:
    post:
      operationId: forceLogoutUser
      tags: [users]
      summary: Force logout all user sessions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /visits:
    get:
      operationId: listVisitRequests
      tags: [visits]
      summary: List visit requests (admin only)
      responses:
        "200":
          description: Visit requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VisitRequest"
    post:
      operationId: createVisitRequest
      tags: [visits]
      summary: Submit a visit request (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitRequestInput"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /visits/{id}/approve:
    post:
      operationId: approveVisitRequest
      tags: [visits]
      summary: Approve a visit request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /visits/{id}/reject:
    post:
      operationId: rejectVisitRequest
      tags: [visits]
      summary: Reject a visit request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        mfaEnabled:
          type: boolean
        preferredLanguage:
          type: string
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
        invitationId:
          type: ["string", "null"]
      required:
        - id
        - fullName
        - email
        - role
        - status
        - mfaEnabled
        - preferredLanguage
        - createdAt
    UserInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string

        role:
          type: string
        password:
          type: string
          minLength: 8
      required:
        - fullName
        - email
        - role
        - password
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
        preferredLanguage:
          type: string
    UserListResult:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - totalPages
...
    VisitRequest:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        reason:
          type: string
        visitorType:
          type: string
        status:
          type: string
        requestedAt:
          type: string
        scheduledAt:
          type: ["string", "null"]
      required:
        - id
        - fullName
        - phone
        - email
        - reason
        - visitorType
        - status
        - requestedAt
    VisitRequestInput:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string

        reason:
          type: string
        visitorType:
          type: string
      required:
        - fullName
        - phone
        - email
        - reason
        - visitorType
    RejectInput:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason